WIP: Removed DocumentImeConnection and related APIs
