+++ super_editor/lib/src/default_editor/document_ime/document_ime_communication.dart
@@ -52,6 +52,10 @@ class DocumentToImeSynchronizer extends StatefulWidget {
   /// this fact.
   final ValueListenable<TextInputConnection?> imeConnection;
 
+  // TODO: the only reason we require a DocumentImeInputClient instead of
+  //       a regular client is so we can set the editing value on it. We should
+  //       either get this working with a regular client, or define a much more
+  //       narrow API
   /// A client that knows how to talk to the platform IME, by receiving
   /// content updates from the platform, and then subsequently sending
   /// new content values to the platform.
@@ -262,260 +266,10 @@ class _DocumentToImeSynchronizerState extends State<DocumentToImeSynchronizer> {
   }
 }
 
-// TODO: get rid of this class
-//
-// This class is a temporary hack so that our synchronization widget doesn't serialize and
-// send documents while we're applying deltas.
-//
-// The root problem here is that we're letting the document send change notifications in
-// the middle of a series of edits. From a public perspective, those edits should be seen
-// as a single edit. The upcoming atomic commands API should make sure to include tooling
-// to prevent change notifications until all pending changes have been applied.
-class ImeEditorStatus {
-  ImeEditorStatus(this._editor);
-
-  final DocumentImeInputClient _editor;
-
-  bool get isEditInProgress => _editor.isApplyingDeltas;
-}
-
-/// A slice of IME functionality that's required by [DocumentToImeSynchronizer].
-abstract class ImeValue {
-  TextEditingValue get currentTextEditingValue;
-
-  set currentTextEditingValue(TextEditingValue newValue);
-
-  bool get isConnectedToIme;
-
-  void closeConnection();
-}
-
-abstract class ImeConnection with ChangeNotifier {
-  /// Returns `true` if there's an open connection to the platform
-  /// Input Method Engine (IME).
-  bool get isAttached;
-
-  /// Opens a connection to the platform Input Method Engine (IME).
-  bool open();
-
-  /// Shows the input controls associated with the IME, e.g., the software keyboard,
-  /// and also connects to the IME, if no connection is present.
-  void show();
-
-  /// Closes the connection to the platform Input Method Engine (IME).
-  void close();
-}
-
-/// Connection to the platform's IME.
-class DocumentImeConnection with ChangeNotifier implements ImeConnection {
-  DocumentImeConnection({
-    required DeltaTextInputClient imeClient,
-  }) {
-    _imeClient = _ClosureAwareImeClientDecorator(imeClient, _onConnectionClosed);
-  }
-
-  late final DeltaTextInputClient _imeClient;
-
-  /// Returns `true` if this [DocumentImeConnection] is currently connected to the platform
-  /// Input Method Engine (IME).
-  ///
-  /// Note, it's possible to have an IME connection with a hidden keyboard. To
-  /// ensure the keyboard is visible for a given IME connection, call [show].
-  @override
-  bool get isAttached => imeConnection != null && imeConnection!.attached;
-
-  TextInputConnection? get imeConnection => _imeConnection;
-  TextInputConnection? _imeConnection;
-  set imeConnection(TextInputConnection? newConnection) {
-    if (newConnection == _imeConnection) {
-      return;
-    }
-
-    editorImeLog.fine("[DocumentImeConnection] - received new IME connection: $newConnection");
-    _imeConnection = newConnection;
-    if (_imeConnection != null) {
-      configureIme(_imeConfig);
-    }
-    notifyListeners();
-  }
-
-  TextInputConfiguration get imeConfiguration => _imeConfig;
-  TextInputConfiguration _imeConfig = const TextInputConfiguration();
-
-  /// Connects the platform Input Method Engine (IME) and (optionally) shows the
-  /// standard input UI, e.g., a software keyboard.
-  ///
-  /// The software keyboard might hide while an IME connection remains open.
-  /// To show the software keyboard for an existing IME connection, use
-  /// [showImeInput].
-  @override
-  bool open({bool showImeInput = true}) {
-    if (isAttached) {
-      // We're already connected to the IME.
-      return true;
-    }
-
-    _imeConnection = TextInput.attach(_imeClient, _imeConfig);
-    editorImeLog.fine("[DocumentImeConnection] - Opened IME connection: $imeConnection");
-    if (imeConnection!.attached == false) {
-      // We failed to connect to the platform IME.
-      editorImeLog.warning("[DocumentImeConnection] - new IME connection is not attached. Null'ing it out");
-      _imeConnection = null;
-      return false;
-    }
-
-    if (showImeInput) {
-      imeConnection!.show();
-    }
-
-    // Notify listeners that a connection was opened.
-    notifyListeners();
-    return true;
-  }
-
-  /// Shows the standard input UI (e.g., software keyboard) for an IME
-  /// connection that's already open.
-  ///
-  /// To open an IME connection, use [open].
-  @override
-  void show() {
-    editorImeLog.info("[DocumentImeConnection] - Showing IME");
-    if (!isAttached) {
-      editorImeLog
-          .info("[DocumentImeConnection] - IME is not yet connected/attached. Opening connection to show keyboard.");
-      open();
-      return;
-    }
-
-    editorImeLog.info("[DocumentImeConnection] - IME is already attached. Calling show()");
-    imeConnection!.show();
-  }
-
-  /// Sets the desired [TextInputConfiguration] for the IME to the given [config],
-  /// and updates the existing platform configuration, if this [DocumentImeConnection] is
-  /// currently connected to the platform IME.
-  void configureIme(TextInputConfiguration config) {
-    if (config == _imeConfig) {
-      return;
-    }
-
-    _imeConfig = config;
-    if (isAttached) {
-      _imeConnection!.updateConfig(_imeConfig);
-    }
-  }
-
-  /// Closes an open connection to the platform's Input Method Engine (IME).
-  @override
-  void close() {
-    if (!isAttached) {
-      return;
-    }
-
-    editorImeLog.info("[DocumentImeConnection] - Closing IME connection");
-    _imeConnection!.close();
-    _imeConnection = null;
-
-    // We notify listeners about closing the connection. Technically, we shouldn't have to
-    // notify listeners here, because the IME client is supposed to get a closure message.
-    // However, we're not getting that message in widget tests. Either Flutter is completely
-    // broken with the message, or Flutter's test IME implementation is broken. To get around
-    // that in tests, we notify listeners from here.
-    notifyListeners();
-  }
-
-  void _onConnectionClosed() {
-    // Notify listeners that the connection was closed.
-    notifyListeners();
-  }
-}
-
-/// A [DeltaTextInputClient] that forwards all calls to the given [_client], and
-/// also notifies [_onConnectionClosed] when the IME connection closes.
-///
-/// This decorator is needed because [TextInputConnection] has no way to listen
-/// for when its connection is closed. By wrapping its [TextInputClient] with
-/// this decorator, the code that owns the [TextInputConnection] can receive
-/// a notification when the connection closes.
-class _ClosureAwareImeClientDecorator implements DeltaTextInputClient {
-  _ClosureAwareImeClientDecorator(this._client, this._onConnectionClosed);
-
-  final DeltaTextInputClient _client;
-  final VoidCallback _onConnectionClosed;
-
-  @override
-  void connectionClosed() {
-    editorImeLog.fine("[_ClosureAwareImeClientDecorator] - IME connection was closed");
-    _onConnectionClosed();
-    _client.connectionClosed();
-  }
-
-  @override
-  AutofillScope? get currentAutofillScope => _client.currentAutofillScope;
-
-  @override
-  TextEditingValue? get currentTextEditingValue => _client.currentTextEditingValue;
-
-  @override
-  void didChangeInputControl(TextInputControl? oldControl, TextInputControl? newControl) {
-    _client.didChangeInputControl(oldControl, newControl);
-  }
-
-  @override
-  void insertTextPlaceholder(Size size) {
-    _client.insertTextPlaceholder(size);
-  }
-
-  @override
-  void performAction(TextInputAction action) {
-    _client.performAction(action);
-  }
-
-  @override
-  void performPrivateCommand(String action, Map<String, dynamic> data) {
-    _client.performPrivateCommand(action, data);
-  }
-
-  @override
-  void performSelector(String selectorName) {
-    _client.performSelector(selectorName);
-  }
-
-  @override
-  void removeTextPlaceholder() {
-    _client.removeTextPlaceholder();
-  }
-
-  @override
-  void showAutocorrectionPromptRect(int start, int end) {
-    _client.showAutocorrectionPromptRect(start, end);
-  }
-
-  @override
-  void showToolbar() {
-    _client.showToolbar();
-  }
-
-  @override
-  void updateEditingValue(TextEditingValue value) {
-    _client.updateEditingValue(value);
-  }
-
-  @override
-  void updateEditingValueWithDeltas(List<TextEditingDelta> textEditingDeltas) {
-    _client.updateEditingValueWithDeltas(textEditingDeltas);
-  }
-
-  @override
-  void updateFloatingCursor(RawFloatingCursorPoint point) {
-    _client.updateFloatingCursor(point);
-  }
-}
-
 /// A [TextInputClient] that applies IME operations to a [Document].
 class DocumentImeInputClient with TextInputClient, DeltaTextInputClient {
   DocumentImeInputClient({
-    // TODO: textDeltasDocumentEditor is enough for text editing deltas, but about
+    // TODO: textDeltasDocumentEditor is enough for text editing deltas, but what about
     //       the non-delta IME operations, like perform selector?
     required this.textDeltasDocumentEditor,
     required this.sendTextEditingValueToIme,
